implement test_take
